Made the player stop when shield it at 0 health or when ball is on players side.
Added required featuers to shieldManager and playerManager to stun player when a shield at 0 health is hit.
Created playerBallStopper to detect and tell the player when the ball is on their side.
Added a new layer that seperates the two sides and only the player can collide with. (Perhaps only the head should specifically collide?)