Attempted to fix new player insantiator code.
in playerManager.cs, at the end of the Start() method is a code to set the joining players to the player spots 1-4.
as a part of this, the code is supposed to set the players onto their appropriate spawn positions, but when I test this with two players, player 2 first spawns on spawn 2, but then gets moved to spawn 1 for some reason.